move "dialog.ts" to ui folder
