+++ core/code/iitc_compability.ts
@@ -1,0 +1,6 @@
+// Export functions for old IITC-Plugins
+import { alert, dialog } from "./ui/dialog";
+
+
+globalThis.alert = alert as any;
+globalThis.dialog = dialog;
